Extra item on IBEX menu to exit, and moved PerspectiveHandler out of package where it was the only class
